+++ src/main/java/com/prography/lighton/artist/application/service/ArtistService.java
@@ -3,8 +3,8 @@ package com.prography.lighton.artist.application.service;
 import com.prography.lighton.artist.domain.entity.Artist;
 import com.prography.lighton.artist.domain.entity.vo.History;
 import com.prography.lighton.artist.infrastructure.repository.ArtistRepository;
-import com.prography.lighton.artist.presentation.dto.RegisterArtistRequest;
-import com.prography.lighton.artist.presentation.dto.UpdateArtistRequest;
+import com.prography.lighton.artist.presentation.dto.request.RegisterArtistRequest;
+import com.prography.lighton.artist.presentation.dto.request.UpdateArtistRequest;
 import com.prography.lighton.common.domain.vo.RegionInfo;
 import com.prography.lighton.genre.application.service.GenreService;
 import com.prography.lighton.genre.domain.entity.Genre;
