+++ src/main/java/com/prography/lighton/artist/presentation/dto/request/RegisterArtistRequest.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.artist.presentation.dto;
+package com.prography.lighton.artist.presentation.dto.request;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
@@ -8,15 +8,19 @@ import jakarta.validation.constraints.Size;
 import java.util.List;
 import org.hibernate.validator.constraints.URL;
 
-public record UpdateArtistRequest(
+public record RegisterArtistRequest(
 
         @NotNull(message = "아티스트 정보는 필수입니다.")
         @Valid
-        RegisterArtistRequest.ArtistDTO artist,
+        ArtistDTO artist,
 
         @NotNull(message = "활동 이력 정보는 필수입니다.")
         @Valid
-        RegisterArtistRequest.HistoryDTO history
+        HistoryDTO history,
+
+        @NotBlank(message = "증빙 자료 URL은 필수입니다.")
+        @URL(message = "증빙 자료는 올바른 URL 형식이어야 합니다.")
+        String proof
 
 ) {
     public record ArtistDTO(
