+++ src/main/java/com/prography/lighton/artist/presentation/dto/request/UpdateArtistRequest.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.artist.presentation.dto;
+package com.prography.lighton.artist.presentation.dto.request;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
@@ -8,19 +8,15 @@ import jakarta.validation.constraints.Size;
 import java.util.List;
 import org.hibernate.validator.constraints.URL;
 
-public record RegisterArtistRequest(
+public record UpdateArtistRequest(
 
         @NotNull(message = "아티스트 정보는 필수입니다.")
         @Valid
-        ArtistDTO artist,
+        RegisterArtistRequest.ArtistDTO artist,
 
         @NotNull(message = "활동 이력 정보는 필수입니다.")
         @Valid
-        HistoryDTO history,
-
-        @NotBlank(message = "증빙 자료 URL은 필수입니다.")
-        @URL(message = "증빙 자료는 올바른 URL 형식이어야 합니다.")
-        String proof
+        RegisterArtistRequest.HistoryDTO history
 
 ) {
     public record ArtistDTO(
